remove plotting code from tree module
