+++ adaptoctree/tree.py
@@ -59,12 +59,12 @@ def build_tree(
         if (level == maximum_level):
             built = True
 
+        # Heavy lifting
         source_keys = morton.encode_points(sources, level, octree_center, octree_radius)
         target_keys = morton.encode_points(targets, level, octree_center, octree_radius)
 
         particle_keys = np.hstack((source_keys, target_keys))
         particle_index_array = np.argsort(particle_keys)
-
         unique_keys, counts = np.unique(particle_keys, return_counts=True) # O(N)
 
         refined_sources = []
@@ -100,75 +100,3 @@ def build_tree(
             targets = np.concatenate(refined_targets)
 
     return tree
-
-
-def plot_tree(tree, octree_center, octree_radius):
-    """
-
-    Parameters:
-    -----------
-    tree : [Node]
-    """
-
-    import itertools
-
-    import matplotlib.pyplot as plt
-    from mpl_toolkits.mplot3d import Axes3D
-
-    points = []
-
-    fig = plt.figure()
-    ax = fig.add_subplot(111, projection='3d')
-
-    unique = []
-
-    for node in tree:
-        level = morton.find_level(node.key)
-        radius = octree_radius / (1 << level)
-
-        center = morton.find_center_from_key(node.key, octree_center, octree_radius)
-
-        r = [-radius, radius]
-
-        for s, e in itertools.combinations(np.array(list(itertools.product(r, r, r))), 2):
-            if np.sum(np.abs(s-e)) == r[1]-r[0]:
-                ax.plot3D(*zip(s+center, e+center), color="b")
-
-    # Plot particle data
-    sources = tree[0].sources
-    ax.scatter(sources[:, 0], sources[:, 1], sources[:, 2], c='g', s=0.8)
-    plt.show()
-
-
-def main():
-    np.random.seed(0)
-
-    def make_moon(npoints):
-
-        x = np.linspace(0, 2*np.pi, npoints) + np.random.rand(npoints)
-        y = 0.5*np.ones(npoints) + np.random.rand(npoints)
-        z = np.sin(x) + np.random.rand(npoints)
-
-        moon = np.array([x, y, z]).T
-        return moon
-
-    sources = targets = make_moon(250)
-
-    tree_conf = {
-        "sources": sources,
-        "targets": targets,
-        "maximum_level": 5,
-        "maximum_particles": 50
-    }
-
-    # Sort sources and targets by octant at level 1 of octree
-    tree = build_tree(**tree_conf)
-
-    max_bound, min_bound = morton.find_bounds(tree_conf['sources'], tree_conf['targets'])
-    octree_center = morton.find_center(max_bound, min_bound)
-    octree_radius = morton.find_radius(octree_center, max_bound, min_bound)
-    plot_tree(tree, octree_center, octree_radius)
-
-
-if __name__ == "__main__":
-    main()
